Add kinematic character movement + assets
RigidBody movement works fine for resim, but it's kind of jank to set up
because there's also an engine callback into the physics_simulate
function we added (which we can't control). Also, requires modifying and
recompiling the source code to expose the simulate function.

Had a hunch that kinematic motion actually works fine, and it was just
the built in character controller that had issues with resim. This
turned out to be true: Godot's Character controller class has some
internal state we can't access, so it's not possible to resimulate
starting from the same conditions.

Decided to scale everything up by about one order of magnitude. Seems
like the collision error magnitude does not scale linearly with size, so
bigger world -> errors seem smaller relatively.

Also added a bunch of assets (including Prism Lad, a player model).
These assets required some import cleaning, so there's a script for
that (MedievalAssets).