Misc character controller + asset improvements
Semi-fix for jitter when both moving and rotating camera. There are two
causes. First, inconsistent frame times mean the rotation amount can
vary a lot between frames, even if you're trying to move mouse smoothly.
FYI, the mouse motion input is sampled once per render frame. Second,
there is an intrinsic rotation jitter in the mouse input. Not entirely
sure if the latter is human error or something fundamental to the mouse
or Godot's input capture. Also not entirely certain if the first one is
really an issue. Fix for the first issue is instead of displaying the
latest rotation, interpolate between the rotation sampled during the
last two physics frames. This should smooth out frametime differences.
Second issue fixed by a second layer of interpolation, just generic
smoothing where the Camera interpolates towards a target rotation.
This does add some latency, but only 2 ish frames, which is not very
noticeable.

Asset transformer can now save assets into sub-folders. It's nice to
have related building blocks grouped together.

Also improved floor snapping. A big issue with the old technique was
that when the player climbs up a ramp onto a flat surface, they'd hover
over the transitional edge, as if it were a low-gravity zone. The reason
is that the old floor snap slowly lowers players to the ground, instead
of instantly snapping. This was done to avoid a noticeable downward jolt
when an airborne player encounters a floor. What we really want is to
have the player stick close to the floor if we're moving between floors
but not if we're airborne. To this end, we add a boolean which tracks
if we're in "floored" mode or "airborne" mode. If airborne, we reduce
the vertical distance of the floor check, so that if we do find a floor,
it's guaranteed that we're already sticking to it. However, while in
floored mode, it's still desirable to do the usual floor check, so that
if cresting a hill-type structure, we will go down the opposite slope
instead of flying away.